+++ src/model/product/DigitalProduct.java
@@ -2,7 +2,7 @@ package model.product;
 
 public class DigitalProduct extends BaseProduct{
 
-    public DigitalProduct(String name, int price, int quantity) {
+    public DigitalProduct(String name, double price, int quantity) {
         super(name, price, quantity);
     }
 }
