+++ src/model/product/ShippableProduct.java
@@ -2,7 +2,7 @@ package model.product;
 
 public class ShippableProduct extends BaseProduct implements IShippableProduct{
     private double weight;
-    public ShippableProduct(String name, int price, int quantity, double weight) {
+    public ShippableProduct(String name, double price, int quantity, double weight) {
         super(name, price, quantity);
         if (weight <= 0){
             throw new IllegalArgumentException("Weight must be greater than 0");
